delete provider factory
